requirements.txt
apache-airflow-providers-apache-spark==5.3.2
apache-airflow-providers-telegram==4.9.0
pandas==2.2.2
pyspark==3.5.2
boto3==1.34.158
clickhouse-connect==0.7.18
unidecode